Simplify trimmable marshaler follow-up
Keep value marshaler helper classes as top-level internals and remove the primitive-array/type-manager helper API churn from the Java.Interop branch. Android keeps the trimmable-only primitive array handling locally.
