+++ src/Java.Interop/PublicAPI.Unshipped.txt
@@ -1,7 +1,6 @@
 #nullable enable
 static Java.Interop.JniEnvironment.BeginMarshalMethod(nint jnienv, out Java.Interop.JniTransition transition, out Java.Interop.JniRuntime? runtime) -> bool
 static Java.Interop.JniEnvironment.EndMarshalMethod(ref Java.Interop.JniTransition transition) -> void
-static Java.Interop.JniRuntime.JniTypeManager.TryGetPrimitiveArrayTypeSignature(System.Type! type, out Java.Interop.JniTypeSignature signature) -> bool
 virtual Java.Interop.JniRuntime.OnEnterMarshalMethod() -> void
 virtual Java.Interop.JniRuntime.OnUserUnhandledException(ref Java.Interop.JniTransition transition, System.Exception! e) -> void
 virtual Java.Interop.JniRuntime.JniTypeManager.GetInvokerTypeCore(System.Type! type) -> System.Type?
@@ -13,7 +12,6 @@ static Java.Interop.JniRuntime.JniTypeManager.TryRegisterBuiltInNativeMembers(Ja
 static Java.Interop.JniRuntime.JniValueManager.CreateDelegatingValueMarshaler<T>(Java.Interop.JniValueMarshaler! valueMarshaler) -> Java.Interop.JniValueMarshaler<T>!
 static Java.Interop.JniRuntime.JniValueManager.ObjectValueMarshaler.get -> Java.Interop.JniValueMarshaler<object?>!
 static Java.Interop.JniRuntime.JniValueManager.PeerableValueMarshaler.get -> Java.Interop.JniValueMarshaler<Java.Interop.IJavaPeerable?>!
-static Java.Interop.JniRuntime.JniValueManager.TryGetPrimitiveArrayValueMarshaler(System.Type! type, out Java.Interop.JniValueMarshaler? marshaler) -> bool
 Java.Interop.JavaException.JavaException(ref Java.Interop.JniObjectReference reference, Java.Interop.JniObjectReferenceOptions transfer, Java.Interop.JniObjectReference throwableOverride) -> void
 Java.Interop.JavaException.SetJavaStackTrace(Java.Interop.JniObjectReference peerReferenceOverride = default(Java.Interop.JniObjectReference)) -> void
 Java.Interop.JniRuntime.ReflectionJniTypeManager
